add booking page

+++ busticket/src/main/java/com/myproject/busticket/repositories/RouteCheckpointRepository.java
@@ -1,12 +1,13 @@
 package com.myproject.busticket.repositories;
 
-import com.myproject.busticket.models.Route_Checkpoint;
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import java.util.List;
+import com.myproject.busticket.models.Route_Checkpoint;
 
-public interface RouteCheckpointRepository extends JpaRepository<Route_Checkpoint, Long> {
+public interface RouteCheckpointRepository extends JpaRepository<Route_Checkpoint, Integer> {
     @Query("select distinct rc.checkpointProvince from Route_Checkpoint rc")
     List<String> findAllCheckpointProvinces();
